Rearrange some other helper functions
